Cambio de tipo de datos y cambio nombre de las variables

+++ core/impuestovehicular-ejb/src/main/java/us/codezen/owner/dto/QueryOwnerReqDTO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @AllArgsConstructor
 @Data
 public class QueryOwnerReqDTO{
-    private int identification;
+    private Long identification;
 }
