+++ core/impuestovehicular-ejb/src/main/java/us/codezen/owner/dto/SearchOwnerReqDTO.java
@@ -6,5 +6,5 @@ import lombok.Data;
 @AllArgsConstructor
 @Data
 public class SearchOwnerReqDTO{
-    private int identification;
+    private Long identification;
 }
